Add an actual test and switch to datetime2
