+++ src/DotJEM.Json.Storage2/SqlServer/Statements/CreateDataTable.sql
@@ -2,8 +2,8 @@
     [Id] [uniqueidentifier] NOT NULL,
     [Version] [int] NOT NULL,
     [ContentType] [varchar](64) NOT NULL,
-    [Created] [datetime] NOT NULL,
-    [Updated] [datetime] NOT NULL,
+    [Created] [datetime2] NOT NULL,
+    [Updated] [datetime2] NOT NULL,
     [CreatedBy] [nvarchar](256) NULL,
     [UpdatedBy] [nvarchar](256) NULL,
     [Data] [nvarchar](max) NOT NULL,
